sandbox: implement more methods of the File interface
