+++ internal/sandbox/dirfs.go
@@ -1,0 +1,3 @@
+package sandbox
+
+func DirFS(path string) FileSystem { return dirFS(path) }
