+++ internal/sandbox/fs.go
@@ -4,8 +4,7 @@ import (
 	"errors"
 	"io/fs"
 	"path"
-	"path/filepath"
-	"strings"
+	"time"
 )
 
 const (
@@ -29,7 +28,7 @@ func OpenRoot(fsys FileSystem) (File, error) {
 }
 
 func Lstat(fsys FileSystem, name string) (fs.FileInfo, error) {
-	info, err := withRoot(fsys, func(dir File) (fs.FileInfo, error) {
+	info, err := atRoot(fsys, func(dir File) (fs.FileInfo, error) {
 		return dir.Lstat(name)
 	})
 	if err != nil {
@@ -39,7 +38,7 @@ func Lstat(fsys FileSystem, name string) (fs.FileInfo, error) {
 }
 
 func Stat(fsys FileSystem, name string) (fs.FileInfo, error) {
-	info, err := withRoot(fsys, func(dir File) (fs.FileInfo, error) {
+	info, err := atRoot(fsys, func(dir File) (fs.FileInfo, error) {
 		for i := 0; i < MaxFollowSymlink; i++ {
 			stat, err := dir.Lstat(name)
 			if err != nil {
@@ -48,7 +47,7 @@ func Stat(fsys FileSystem, name string) (fs.FileInfo, error) {
 			if stat.Mode().Type() != fs.ModeSymlink {
 				return stat, nil
 			}
-			link, err := dir.ReadLink(name)
+			link, err := dir.Readlink(name)
 			if err != nil {
 				return nil, err
 			}
@@ -65,7 +64,7 @@ func Stat(fsys FileSystem, name string) (fs.FileInfo, error) {
 	return info, nil
 }
 
-func withRoot[F func(File) (R, error), R any](fsys FileSystem, do F) (ret R, err error) {
+func atRoot[F func(File) (R, error), R any](fsys FileSystem, do F) (ret R, err error) {
 	d, err := OpenRoot(fsys)
 	if err != nil {
 		return ret, err
@@ -99,21 +98,29 @@ type File interface {
 
 	Datasync() error
 
+	Truncate(size int64) error
+
+	SetFlags(flags int) error
+
 	ReadDir(n int) ([]fs.DirEntry, error)
 
 	Lstat(name string) (fs.FileInfo, error)
 
-	ReadLink(name string) (string, error)
+	Readlink(name string) (string, error)
 
-	SetFlags(flags int) error
-}
+	Chtimes(name string, atime, mtime time.Time) error
 
-func DirFS(path string) (FileSystem, error) {
-	path, err := filepath.Abs(path)
-	if err != nil {
-		return nil, err
-	}
-	return dirFS(path), nil
+	Mkdir(name string, mode uint32) error
+
+	Rmdir(name string) error
+
+	Rename(oldName string, newDir File, newName string) error
+
+	Link(oldName string, newDir File, newName string) error
+
+	Symlink(oldName, newName string) error
+
+	Unlink(name string) error
 }
 
 func FS(fsys FileSystem) fs.FS {
@@ -139,9 +146,9 @@ func (fsys *fsFileSystem) Stat(name string) (fs.FileInfo, error) {
 	}
 	s, err := Stat(fsys.base, name)
 	if err != nil {
-		err = fsError("stat", name, err)
+		return nil, fsError("stat", name, err)
 	}
-	return s, err
+	return s, nil
 }
 
 func fsError(op, name string, err error) error {
@@ -157,6 +164,14 @@ func fsError(op, name string, err error) error {
 		err = fs.ErrClosed
 	}
 	return &fs.PathError{Op: op, Path: name, Err: err}
+}
+
+func unwrapPathError(err error) error {
+	e, ok := err.(*fs.PathError)
+	if ok {
+		return e.Err
+	}
+	return err
 }
 
 var (
@@ -193,228 +208,4 @@ func (dirent *fsDirEntry) Info() (fs.FileInfo, error) {
 		return nil, fsError("stat", dirent.Name(), err)
 	}
 	return s, nil
-}
-
-func RootFS(fsys FileSystem) FileSystem {
-	return &rootFS{fsys}
-}
-
-type rootFS struct{ base FileSystem }
-
-func (fsys *rootFS) Open(name string, flags int, mode fs.FileMode) (File, error) {
-	f, err := OpenRoot(fsys.base)
-	if err != nil {
-		return nil, err
-	}
-	defer f.Close()
-	return (&rootFile{f}).Open(name, flags, mode)
-}
-
-type nopFileCloser struct{ File }
-
-func (nopFileCloser) Close() error { return nil }
-
-type rootFile struct{ File }
-
-func (f *rootFile) Open(name string, flags int, mode fs.FileMode) (File, error) {
-	file, err := f.open(name, flags, mode)
-	if err != nil {
-		return nil, &fs.PathError{Op: "open", Path: name, Err: unwrapPathError(err)}
-	}
-	return &rootFile{file}, nil
-}
-
-func (f *rootFile) open(name string, flags int, mode fs.FileMode) (File, error) {
-	dir := File(nopFileCloser{f.File})
-	defer func() { dir.Close() }()
-	setCurrentDirectory := func(cd File) {
-		dir.Close()
-		dir = cd
-	}
-
-	followSymlinkDepth := 0
-	followSymlink := func(symlink, target string) error {
-		link, err := dir.ReadLink(symlink)
-		if err != nil {
-			// This error may be EINVAL if the file system was modified
-			// concurrently and the directory entry was not pointing to a
-			// symbolic link anymore.
-			return err
-		}
-
-		// Limit the maximum number of symbolic links that would be followed
-		// during path resolution; this ensures that if we encounter a loop,
-		// we will eventually abort resolving the path.
-		if followSymlinkDepth == MaxFollowSymlink {
-			return ELOOP
-		}
-		followSymlinkDepth++
-
-		if target != "" {
-			name = link + "/" + target
-		} else {
-			name = link
-		}
-		return nil
-	}
-
-	depth := filePathDepth(f.Name())
-	for {
-		if isAbs(name) {
-			name = trimLeadingSlash(name)
-			d, err := f.openRoot()
-			if err != nil {
-				return nil, err
-			}
-			if name == "" {
-				return d, nil
-			}
-			depth = 0
-			setCurrentDirectory(d)
-		}
-
-		var move int
-		var elem string
-		elem, name = splitFilePath(name)
-
-		switch elem {
-		case ".":
-		openFile:
-			newFile, err := dir.Open(name, flags|O_NOFOLLOW, mode)
-			if err != nil {
-				if !errors.Is(err, ELOOP) || ((flags & O_NOFOLLOW) != 0) {
-					return nil, err
-				}
-				switch err := followSymlink(name, ""); err {
-				case nil:
-					continue
-				case EINVAL:
-					goto openFile
-				default:
-					return nil, err
-				}
-			}
-			return newFile, nil
-
-		case "..":
-			// This check ensures that we cannot escape the root of the file
-			// system when accessing a parent directory.
-			if depth == 0 {
-				continue
-			}
-			move = -1
-		default:
-			move = +1
-		}
-
-	openPath:
-		d, err := dir.Open(elem, openPathFlags, 0)
-		if err != nil {
-			if !errors.Is(err, ENOTDIR) {
-				return nil, err
-			}
-			switch err := followSymlink(elem, name); err {
-			case nil:
-				continue
-			case EINVAL:
-				goto openPath
-			default:
-				return nil, err
-			}
-		}
-		depth += move
-		setCurrentDirectory(d)
-	}
-}
-
-func (f *rootFile) openRoot() (File, error) {
-	depth := filePathDepth(f.Name())
-	if depth == 0 {
-		return f.Open(".", O_DIRECTORY, 0)
-	}
-	dir := File(nopFileCloser{f.File})
-	for depth > 0 {
-		p, err := dir.Open("..", O_DIRECTORY, 0)
-		if err != nil {
-			return nil, err
-		}
-		dir.Close()
-		dir = p
-	}
-	return dir, nil
-}
-
-func (f *rootFile) Lstat(name string) (fs.FileInfo, error) {
-	return withPath(f, "stat", name, File.Lstat)
-}
-
-func (f *rootFile) ReadLink(name string) (string, error) {
-	return withPath(f, "readlink", name, File.ReadLink)
-}
-
-func withPath[F func(File, string) (R, error), R any](root *rootFile, op, name string, do F) (ret R, err error) {
-	dir, base := path.Split(name)
-	if dir == "" {
-		return do(root.File, base)
-	}
-	d, err := root.Open(dir, openPathFlags, 0)
-	if err != nil {
-		return ret, &fs.PathError{Op: op, Path: name, Err: unwrapPathError(err)}
-	}
-	defer d.Close()
-	return do(d.(*rootFile).File, base)
-}
-
-func unwrapPathError(err error) error {
-	e, ok := err.(*fs.PathError)
-	if ok {
-		return e.Err
-	}
-	return err
-}
-
-func filePathDepth(path string) (depth int) {
-	for {
-		path = trimLeadingSlash(path)
-		if path == "" {
-			return depth
-		}
-		depth++
-		i := strings.IndexByte(path, '/')
-		if i < 0 {
-			return depth
-		}
-		path = path[i:]
-	}
-}
-
-func splitFilePath(path string) (elem, name string) {
-	path = trimLeadingSlash(path)
-	path = trimTrailingSlash(path)
-	i := strings.IndexByte(path, '/')
-	if i < 0 {
-		return ".", path
-	} else {
-		return path[:i], trimLeadingSlash(path[i:])
-	}
-}
-
-func trimLeadingSlash(s string) string {
-	i := 0
-	for i < len(s) && s[i] == '/' {
-		i++
-	}
-	return s[i:]
-}
-
-func trimTrailingSlash(s string) string {
-	i := len(s)
-	for i > 0 && s[i-1] == '/' {
-		i--
-	}
-	return s[:i]
-}
-
-func isAbs(path string) bool {
-	return len(path) > 0 && path[0] == '/'
 }
